% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_log.R
\name{reg_log_hd}
\alias{reg_log_hd}
\title{Fits MSE-minimizing logistic regression model}
\usage{
reg_log_hd(X, Y, estimate_gamma = FALSE)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response variable}

\item{estimate_gamma}{if TRUE, signal strength will be estimated from data}
}
\description{
give a deeper explanation
}
\examples{


}
